+++ src/LoginFormContainer.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+export default function LoginFormContainer({ onChange }) {
+  return (
+    <>
+      <div>
+        <label htmlFor="login-email">
+          E-mail
+        </label>
+        <input type="email" id="login-email" onChange={onChange} />
+      </div>
+      <div>
+        <label htmlFor="login-password">
+          Password
+        </label>
+        <input type="password" id="login-password" />
+      </div>
+    </>
+  );
+}
