+++ src/LoginFormContainer.test.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+import { render, fireEvent } from '@testing-library/react';
+
+import LoginFormContainer from './LoginFormContainer';
+
+describe('LoginFormContainer', () => {
+  const handleChange = jest.fn();
+
+  beforeEach(() => {
+    handleChange.mockClear();
+  });
+
+  function renderLoginFormContainer() {
+    return render((
+      <LoginFormContainer onChange={handleChange} />
+    ));
+  }
+
+  it('renders the login form', () => {
+    const { getByLabelText } = renderLoginFormContainer();
+
+    expect(getByLabelText('E-mail')).not.toBeNull();
+    expect(getByLabelText('Password')).not.toBeNull();
+  });
+
+  it('listens form fields change event', () => {
+    const { getByLabelText } = renderLoginFormContainer();
+
+    fireEvent.change(getByLabelText('E-mail'), {
+      target: { value: 'tester@example.com' },
+    });
+
+    expect(handleChange).toBeCalled();
+  });
+});
